+++ CHANGELOG.md
@@ -5,8 +5,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## \[Unreleased\]
 
 ### Added
+
+- Added `revokeZomeCallCapability` to the admin websocket to revoke zome call capabilities for an agent [#360](https://github.com/holochain/holochain-client-js/pull/360)
+- Added `listCapabilityGrants` to the admin websocket to list all zome call capabilities granted to an agent [#360](https://github.com/holochain/holochain-client-js/pull/360)
+
 ### Fixed
+
+- Fixed types for `list_apps` to match the new API
+
 ### Changed
+
+- Changed return type of `grantZomeCallCapability` to `ActionHash` [#360](https://github.com/holochain/holochain-client-js/pull/360)
+
 ### Removed
 
 ## 2025-07-02: v0.20.0-dev.1